pyproject.toml
[tool.poetry]
name = "give-me-insights-backend"
version = "0.1.20"
description = ""
authors = ["JangasCodingplace <janisgoesser92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.13.0"
boto3 = "^1.26.26"
django-storages = "^1.13.1"
gunicorn = "^20.1.0"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
